requirements.txt
django==5.1.3
djangorestframework
django-cors-headers
django-filter